test(shortener): add `integration.test` to validate all `shortener` endpoints
